Change signature for getting data on Virtual Disk and add some asserts.

+++ src/emulator/src/VirtualDisk.h
@@ -17,7 +17,7 @@ public:
 
     VirtualDisk(std::span<std::uint8_t> data, Layout layout);
 
-    [[nodiscard]] std::uint8_t get(int track, int sector, int index_in_sector) const;
+    [[nodiscard]] std::uint8_t get(uint8_t track, uint8_t sector, uint8_t index_in_sector) const;
 
 private:
     std::vector<uint8_t> data;
@@ -28,7 +28,8 @@ private:
     [[nodiscard]] std::uint8_t get_checksum(int track, int sector) const;
     uint8_t computer_sector_checksum(int track, int sector);
 
-    [[nodiscard]] std::uint8_t get_data(int track, int sector, int index_in_sector) const;
+    [[nodiscard]] std::uint8_t get_data(uint8_t track, uint8_t sector,
+                                        uint8_t index_in_sector) const;
 
 };
